+++ de.metas.procurement.base/src/main/java/de/metas/procurement/base/model/interceptor/C_BPartner.java
@@ -2,10 +2,12 @@ package de.metas.procurement.base.model.interceptor;
 
 import org.adempiere.ad.modelvalidator.annotations.Interceptor;
 import org.adempiere.ad.modelvalidator.annotations.ModelChange;
+import org.adempiere.exceptions.AdempiereException;
 import org.adempiere.util.Services;
 import org.compiere.model.I_C_BPartner;
 import org.compiere.model.ModelValidator;
 
+import de.metas.procurement.base.IPMMContractsDAO;
 import de.metas.procurement.base.IWebuiPush;
 
 /*
@@ -18,12 +20,12 @@ import de.metas.procurement.base.IWebuiPush;
  * it under the terms of the GNU General Public License as
  * published by the Free Software Foundation, either version 2 of the
  * License, or (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public
  * License along with this program.  If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
@@ -39,11 +41,34 @@ public class C_BPartner
 	{
 	}
 
+	/**
+	 * Make sure an partner with currently running contracts always has <code>IsVendor='Y'</code>.
+	 *
+	 * @param bpartner
+	 */
+	@ModelChange(timings = ModelValidator.TYPE_BEFORE_CHANGE, ifColumnsChanged = I_C_BPartner.COLUMNNAME_IsVendor)
+	public void preventDisablingVendorFlag(final I_C_BPartner bpartner)
+	{
+		// Do nothing if vendor flag is set
+		if (bpartner.isVendor())
+		{
+			return;
+		}
+
+		if (Services.get(IPMMContractsDAO.class).hasRunningContract(bpartner))
+		{
+			throw new AdempiereException("@C_BPartner_ID@ @IsVendor@=@N@: " + bpartner.getValue() + "_" + bpartner.getName());
+		}
+	}
+
 	@ModelChange(timings = ModelValidator.TYPE_AFTER_CHANGE //
-	, ifColumnsChanged = { I_C_BPartner.COLUMNNAME_Name, I_C_BPartner.COLUMNNAME_AD_Language, I_C_BPartner.COLUMNNAME_IsVendor } //
-	, afterCommit = true)
+	, ifColumnsChanged = {
+			I_C_BPartner.COLUMNNAME_Name,
+			I_C_BPartner.COLUMNNAME_AD_Language,
+			I_C_BPartner.COLUMNNAME_IsVendor } //
+			, afterCommit = true)
 	public void pushToWebUI(final I_C_BPartner bpartner)
 	{
-		Services.get(IWebuiPush.class).pushBPartnerWithoutContracts(bpartner);
+		Services.get(IWebuiPush.class).pushBPartnerAndUsers(bpartner);
 	}
 }